sched.cc: Reenable idle polling.
Upstream OSv code polls the runqueue in the idle loop before sleeping
the CPU.  This feature improves responsiveness, but consumes noticably
more CPU than just sleeping immediately in the idle loop.  This change
reenables polling but lowers the number of polling iterations so that
idle CPU usage is not significantly affected.  For reference, the OSv
codebase originally used 100 iterations here, then changed it to
10000.  This change sets the count to an itermediate value, 1000.