memory: Prevent array overflow in l1 page buffer pool.
Under high levels of memory page churn, the array of pages in the
l1 page buffer pool could overflow and overwrite the data in adjacent
memory pages.  This change replaces the array with a fixed sized ring.

It also removes the looping behavior found in the refill and unfill
functions.  Those functions are called by multiple threads and now
only do the minimum amount of work necessary to allow the callers to
continue.  Previously, the looping behavior in the refill thread caused
the array overflow due to a race on the array size.

The pool fill thread is now solely responsible for maintaining
the target pool size.