memory: respect _bitset length bounds
The page_range_allocator attempts to merge neighbor page ranges
when a page range is freed in order to minimize memory fragmentation.

However, the original implementation didn't validate the index before
checking the _bitset to determine whether a page range had a neighbor
or not.  This change make sure the index stays within bounds.

This is the correct fix for the fix attempted in
a29b4749b1839197109bb96d7eb1c4a93082cc1c.