tcp_input.cc: acknowledge SYN 'byte' when connection established
When tcp_do_segment transitions from TCPS_SYN_RECEIVED to TCPS_ESTABLISHED,
the incoming ACK for the outgoing SYN is not accounted for before normal
ACK processing takes place.  As a result, if there is data already in
the socket send buffer when this transition occurs, then the first
byte of the buffer can be trimmed to account for the acknowledged SYN
'byte'.

This change updates the unacknowledged send count (snd_una) when the
transition occurs.