Added NETLINK socket to allow getifaddrs() to support IPv6
The NETLINK socket support is only enough to support getifaddrs().
The code for getifaddrs.c and if_nameindex.c are from the latest
musl code repository.

The only way to get IPv6 interface info on FreeBSD is with sysctl,
but that isn't supported on OSv, so Linux NETLINK socket API
was added to do it.